testing matias

+++ src/context/context.tsx
@@ -102,7 +102,7 @@ export const ProductContextProvider = ({ children }) => {
 
       const { error, data } = await supabase.from("products").insert({
         ...product,
-        user: user.id,
+        userId: user.id,
       });
 
       // setProducts([...products, ...data]);
@@ -110,6 +110,7 @@ export const ProductContextProvider = ({ children }) => {
       if (error) {
         throw error;
       }
+      console.log('data create', data)
     } catch (error) {
       alert(error.error_description || error.message);
     } finally {
@@ -127,7 +128,7 @@ export const ProductContextProvider = ({ children }) => {
         .from("products")
         .select("id, title,description,price,category, image, boost")
 
-        .eq("user", user.id);
+        .eq("userId", user.id);
 
       // .order("id", { ascending: false });
 
@@ -151,11 +152,12 @@ export const ProductContextProvider = ({ children }) => {
       const { error, data } = await supabase
         .from("products")
         .update(updatedFields)
-        .eq("user", user.id)
+        .eq("userId", user.id)
         .eq("id", id);
       if (error) {
         throw error;
       }
+      console.log("update", data)
 
       setProducts(products.filter((Product) => Product.id !== data[0].id));
     } catch (error) {
@@ -171,7 +173,7 @@ export const ProductContextProvider = ({ children }) => {
       const { error, data } = await supabase
         .from("products")
         .delete()
-        .eq("user", user.id)
+        .eq("userId", user.id)
         .eq("id", product.id);
 
       if (error) {
